clean class properties, already defined from Iface
Really, if when we create a content-type,
we are providing and Interface/schema for it,
doesn't make sense to have the same declared as class properties.

Class properties, in general should be avoided, because they belong
to the class not to the instance, and it's easy to got into
so strange behaviors, when you have mutable objects on them,
and they remember the state across calls.